Ensure consistent matching between DOM/composite components
The context of a query could incorrectly exclude a Composite component
from the results if it had been previously filtered by a DOM match.
This would happen because `rquery_findAllInRenderedTree` would only
return DOM descendants when given a DOM node. However, consider the
React tree: <div><MyComponent /></div>. The query `div MyComponent`
would fail because `rquery_findAllInRenderedTree` would be called with
the <div> node, and the `MyComponent` instance would not be returned,
only its related DOM component.

The child selector (>) actually handled this properly by doing a search
on the original root component, which is always a composite component
and allows full traversal of DOM and composite components. The fix is
to include this logic in the descendant selector.